requirements.txt
google-adk
google-cloud-aiplatform
google-genai
google-generativeai
pandas
asyncio
python-dotenv